+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # RELEASE HISTORY
 
+## 4.0.18 November 13, 2017
+* The loginWithTwitterSDK sync and async methods added to the UserService class
+
+## 4.0.17 November 8, 2017
+* Fixed the logout bug when "Users" table is mapped to the BackendlessUser class
+
 ## 4.0.16 October 6, 2017
 * The deviceRegistration (sync/async) methods bug with the default channel fixed
 * The BackendlessPushHelper class created to process mutable content